requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
websockets==12.0
hypersync==0.1.8
pydantic==2.5.0
strenum==0.4.15